Mod: 本番環境用のGitHub API クライアントなどの実装
